Refactor LoadMethod.LOCAL to use symlinks instead of copying directory (#660)
This PR refactors the `create_symlinks` function that was previously
used in load via dbt ls so that it can be used in
`DbtLocalBaseOperator.run_command` instead of copying the entire
directory.

Closes: #614
(cherry picked from commit 5d23758da43ad93e4f6f80a063ff1c76cefe1557)
