+++ dev/dags/dbt/simple/profiles.yml
@@ -7,5 +7,5 @@ simple:
       database: 'database'
       schema: 'main'
       schemas_and_paths:
-        main: "{{ env_var('DBT_SQLITE_PATH', '.') }}/data/imdb.db"
-      schema_directory: 'data'
+        main: "{{ env_var('DBT_SQLITE_PATH') }}/imdb.db"
+      schema_directory: "{{ env_var('DBT_SQLITE_PATH') }}"
